requirements.txt
numpy>=1.21.0
Pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
faiss-cpu>=1.7.0
Flask>=2.0.0
PyYAML>=6.0
tqdm>=4.65.0
pytest>=7.0.0